backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
python-multipart==0.0.9
torch==2.3.0
torchvision==0.18.0
numpy==1.26.4
opencv-python-headless==4.9.0.80
Pillow==10.3.0
safetensors==0.4.3
transformers==4.41.2
sentencepiece==0.2.0
gtts==2.5.1
langdetect==1.0.9
deep-translator==1.11.4
pydantic==2.7.1
